Introduced configuration
Configuration is read from a .gitaccount file in the projects path first. If not found it's loaded from the users home. If not found a default config is used. In addition the config can be set lazily. Every needed option that is not specified in the file is taken from the default config.