Alternative method for current branch and 1.9.2
Some projects don't want to return the current branch in the expected way. There is a new method alt_name that returns the name in an alternate way. 

Second the configuration was not compatible to 1.9.2 cause the dynamic attr_acessors were private. I made the generator public so the methods are. Not nice but works.