Refactored BatchSender
Batches are sent in a completely new fashion now. The commits do not include project data during batch send. Instead the data to be send includes a new hash key :projects that transfers the project data once. According to that change a lot of models were changed, tests added and a bug fixed.